rename: user와 관련된 테스트 파일의 위치를 이동합니다
